feat(app): full new-system extraction (Phase 3b)
Replace the compat-shim hybrid (commit 1cc4380) with full-extension
extraction. App.tsx now contains zero JSX route trees, no convertLegacyApp
calls, and no @backstage/app-defaults references — every plugin and custom
behavior is registered as a new-system FrontendFeature.

Plugin imports:
  All plugins now imported from /alpha (catalog, catalog-graph,
  catalog-import, scaffolder, techdocs, search, api-docs, org,
  user-settings, kubernetes, github-actions). Each /alpha default export
  is a FrontendPlugin that provides its own pages, nav items, and entity
  page contributions.

Custom extensions (packages/app/src/extensions/):
  - apis.ts         scmIntegrationsApi, scmAuthApi, gleanAnalyticsApi
                    via ApiBlueprint
  - signInPage.tsx  SignInPageBlueprint with the dev-vs-prod
                    guest/gcpIap switch preserved
  - sidebar.tsx     NavContentBlueprint with the Mozilla logo, search
                    modal, menu items, and settings group
  - rootElements.tsx  AppRootElementBlueprint for AlertDisplay and
                    OAuthRequestDialog

App.tsx wires it all together via createFrontendModule({ pluginId: 'app' })
and createApp's bindRoutes (using the alpha plugins' externalRoutes).

Files removed (no longer needed — alpha plugins provide defaults):
  - packages/app/src/apis.ts
  - packages/app/src/components/catalog/EntityPage.tsx (415 lines of
    stock create-app boilerplate)
  - packages/app/src/components/Root/Root.tsx, index.ts (replaced by
    sidebar.tsx extension)
  - packages/app/src/components/search/SearchPage.tsx (the alpha search
    plugin provides its own page; this app's customization was minor
    and can be re-added as a search frontend module if needed)

Dependency adjustments:
  - @backstage/frontend-defaults  ^0.2.2 -> ^0.5.1  (was on the old
    line whose nested @backstage/frontend-plugin-api 0.10.x conflicted
    with the alpha plugins' 0.16.x types)
  - @backstage/core-compat-api    ^0.4.2 -> ^0.5.10
  - @backstage-community/plugin-github-actions  ^0.11.1 -> ^1.0.0  (the
    old version pinned @backstage/frontend-plugin-api ^0.10.2, which
    fragmented the type system across nested node_modules)
  - @backstage/plugin-app-react  added (used by SignInPageBlueprint /
    NavContentBlueprint)
  - @backstage/app-defaults      removed

Type system note:
  The Glean analytics class implements AnalyticsApi from
  @backstage/core-plugin-api (legacy), but the new system's
  analyticsApiRef expects AnalyticsApi from @backstage/frontend-plugin-api.
  Runtime shape is compatible (captureEvent(event)); only AnalyticsContextValue
  has stricter fields in the new system. Cast at the registration site;
  Phase 4 (Glean modernization) will resolve this properly.

Verification:
  yarn tsc                        clean
  yarn lint:all                   clean
  yarn prettier:check             clean
  yarn test:all                   8/8 pass
  yarn workspace app build        succeeds (1.4 MiB main bundle)

Net diff: 141 insertions / 1304 deletions across 8 files.

Runtime behavior verification (Postgres + browser) deferred to CI.
