+++ packages/app/package.json
@@ -14,17 +14,18 @@
     "lint": "backstage-cli package lint"
   },
   "dependencies": {
-    "@backstage-community/plugin-github-actions": "^0.11.1",
+    "@backstage-community/plugin-github-actions": "^1.0.0",
     "@backstage/catalog-model": "^1.8.0",
     "@backstage/cli": "^0.36.1",
     "@backstage/core-app-api": "^1.20.0",
-    "@backstage/core-compat-api": "^0.4.2",
+    "@backstage/core-compat-api": "^0.5.10",
     "@backstage/core-components": "^0.18.9",
     "@backstage/core-plugin-api": "^1.12.5",
-    "@backstage/frontend-defaults": "^0.2.2",
+    "@backstage/frontend-defaults": "^0.5.1",
     "@backstage/frontend-plugin-api": "^0.16.0",
     "@backstage/integration-react": "^1.2.17",
     "@backstage/plugin-api-docs": "^0.14.0",
+    "@backstage/plugin-app-react": "^0.2.2",
     "@backstage/plugin-catalog": "^2.0.4",
     "@backstage/plugin-catalog-common": "^1.1.9",
     "@backstage/plugin-catalog-graph": "^0.6.3",
