+++ packages/app/src/App.tsx
@@ -1,112 +1,65 @@
-import { Navigate, Route } from 'react-router-dom';
-import { apiDocsPlugin, ApiExplorerPage } from '@backstage/plugin-api-docs';
-import {
-  CatalogEntityPage,
-  CatalogIndexPage,
-  catalogPlugin,
-} from '@backstage/plugin-catalog';
-import {
-  CatalogImportPage,
-  catalogImportPlugin,
-} from '@backstage/plugin-catalog-import';
-import { ScaffolderPage, scaffolderPlugin } from '@backstage/plugin-scaffolder';
-import { orgPlugin } from '@backstage/plugin-org';
-import { SearchPage } from '@backstage/plugin-search';
-import {
-  TechDocsIndexPage,
-  techdocsPlugin,
-  TechDocsReaderPage,
-} from '@backstage/plugin-techdocs';
-import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
-import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
-import { UserSettingsPage } from '@backstage/plugin-user-settings';
-import { apis } from './apis';
-import { entityPage } from './components/catalog/EntityPage';
-import { searchPage } from './components/search/SearchPage';
-import { Root } from './components/Root';
-
-import {
-  AlertDisplay,
-  OAuthRequestDialog,
-  ProxiedSignInPage,
-  SignInPage,
-} from '@backstage/core-components';
 import { createApp } from '@backstage/frontend-defaults';
 import {
-  convertLegacyApp,
-  convertLegacyAppOptions,
-} from '@backstage/core-compat-api';
-import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
-import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
-import { RequirePermission } from '@backstage/plugin-permission-react';
-import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { configApiRef, useApi } from '@backstage/core-plugin-api';
+  createFrontendModule,
+  FrontendFeature,
+} from '@backstage/frontend-plugin-api';
 
-const routes = (
-  <FlatRoutes>
-    <Route path="/" element={<Navigate to="catalog" />} />
-    <Route path="/catalog" element={<CatalogIndexPage />} />
-    <Route
-      path="/catalog/:namespace/:kind/:name"
-      element={<CatalogEntityPage />}
-    >
-      {entityPage}
-    </Route>
-    <Route path="/docs" element={<TechDocsIndexPage />} />
-    <Route
-      path="/docs/:namespace/:kind/:name/*"
-      element={<TechDocsReaderPage />}
-    >
-      <TechDocsAddons>
-        <ReportIssue />
-      </TechDocsAddons>
-    </Route>
-    <Route path="/create" element={<ScaffolderPage />} />
-    <Route path="/api-docs" element={<ApiExplorerPage />} />
-    <Route
-      path="/catalog-import"
-      element={
-        <RequirePermission permission={catalogEntityCreatePermission}>
-          <CatalogImportPage />
-        </RequirePermission>
-      }
-    />
-    <Route path="/search" element={<SearchPage />}>
-      {searchPage}
-    </Route>
-    <Route path="/settings" element={<UserSettingsPage />} />
-    <Route path="/catalog-graph" element={<CatalogGraphPage />} />
-  </FlatRoutes>
-);
+import catalogPlugin from '@backstage/plugin-catalog/alpha';
+import catalogGraphPlugin from '@backstage/plugin-catalog-graph/alpha';
+import catalogImportPlugin from '@backstage/plugin-catalog-import/alpha';
+import scaffolderPlugin from '@backstage/plugin-scaffolder/alpha';
+import techdocsPlugin from '@backstage/plugin-techdocs/alpha';
+import searchPlugin from '@backstage/plugin-search/alpha';
+import apiDocsPlugin from '@backstage/plugin-api-docs/alpha';
+import orgPlugin from '@backstage/plugin-org/alpha';
+import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
+import kubernetesPlugin from '@backstage/plugin-kubernetes/alpha';
+import githubActionsPlugin from '@backstage-community/plugin-github-actions/alpha';
+import { techDocsReportIssueAddonModule } from '@backstage/plugin-techdocs-module-addons-contrib/alpha';
 
-const rootElement = (
-  <>
-    <AlertDisplay />
-    <OAuthRequestDialog />
-    <AppRouter>
-      <Root>{routes}</Root>
-    </AppRouter>
-  </>
-);
+import {
+  scmIntegrationsApi,
+  scmAuthApi,
+  gleanAnalyticsApi,
+} from './extensions/apis';
+import { signInPageExtension } from './extensions/signInPage';
+import { sidebarExtension } from './extensions/sidebar';
+import {
+  alertDisplayElement,
+  oauthRequestDialogElement,
+} from './extensions/rootElements';
+
+const appModule = createFrontendModule({
+  pluginId: 'app',
+  extensions: [
+    scmIntegrationsApi,
+    scmAuthApi,
+    gleanAnalyticsApi,
+    signInPageExtension,
+    sidebarExtension,
+    alertDisplayElement,
+    oauthRequestDialogElement,
+  ],
+});
+
+const features: FrontendFeature[] = [
+  appModule,
+  catalogPlugin,
+  catalogGraphPlugin,
+  catalogImportPlugin,
+  scaffolderPlugin,
+  techdocsPlugin,
+  techDocsReportIssueAddonModule,
+  searchPlugin,
+  apiDocsPlugin,
+  orgPlugin,
+  userSettingsPlugin,
+  kubernetesPlugin,
+  githubActionsPlugin,
+];
 
 const app = createApp({
-  features: [
-    convertLegacyAppOptions({
-      apis,
-      components: {
-        SignInPage: props => {
-          const configApi = useApi(configApiRef);
-          if (
-            configApi.getOptionalString('auth.environment') === 'development'
-          ) {
-            return <SignInPage {...props} providers={['guest']} />;
-          }
-          return <ProxiedSignInPage {...props} provider="gcpIap" />;
-        },
-      },
-    }),
-    ...convertLegacyApp(rootElement),
-  ],
+  features,
   bindRoutes({ bind }) {
     bind(catalogPlugin.externalRoutes, {
       createComponent: scaffolderPlugin.routes.root,
