--- packages/app/src/components/Root/index.ts
@@ -1,1 +1,0 @@
-export { Root } from './Root';
